Add tests for actions
